Extract config classes from main.py to config.py
Move StepMeanEntry, InferenceEvaluatorAggregatorConfig, and
InferenceAggregatorConfig to a dedicated config module to
decouple configuration from aggregator implementation.
